@@ -1,6 +1,6 @@
 module github.com/mendersoftware/mender-cli
 
-go 1.18
+go 1.23.0
 
 require (
 	github.com/cheggaaa/pb/v3 v3.1.7
@@ -14,8 +14,8 @@ require (
 	github.com/spf13/pflag v1.0.6
 	github.com/spf13/viper v1.19.0
 	github.com/vmihailenco/msgpack v4.0.4+incompatible
-	golang.org/x/sys v0.30.0
-	golang.org/x/term v0.29.0
+	golang.org/x/sys v0.31.0
+	golang.org/x/term v0.30.0
 )
 
 require (
